algunas modificaciones y correcciones
